Organized source code files, fixed lagging options state in expanded multiple choice item
